+++ modules/global/src/org/mg/mgweb/entity/TblTransferencias.java
@@ -53,10 +53,10 @@ public class TblTransferencias extends BaseIntIdentityIdEntity {
     protected com.haulmont.cuba.security.entity.User idUser;
 
     @Column(name = "monto_aporte", nullable = false)
-    protected Integer montoAporte;
+    protected Long montoAporte;
 
     @Column(name = "monto_donacion", nullable = false)
-    protected Integer montoDonacion;
+    protected Long montoDonacion;
 
     @Column(name = "seq_pago", nullable = false)
     protected Integer seqPago;
@@ -83,19 +83,19 @@ public class TblTransferencias extends BaseIntIdentityIdEntity {
         this.seqPago = seqPago;
     }
 
-    public Integer getMontoDonacion() {
+    public Long getMontoDonacion() {
         return montoDonacion;
     }
 
-    public void setMontoDonacion(Integer montoDonacion) {
+    public void setMontoDonacion(Long montoDonacion) {
         this.montoDonacion = montoDonacion;
     }
 
-    public Integer getMontoAporte() {
+    public Long getMontoAporte() {
         return montoAporte;
     }
 
-    public void setMontoAporte(Integer montoAporte) {
+    public void setMontoAporte(Long montoAporte) {
         this.montoAporte = montoAporte;
     }
